requirements.txt
Flask==2.3.3
rembg==2.0.50
Pillow==10.2.0
numpy==1.26.4